features/requirements.txt
polars>=0.20.0
numpy>=1.24.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
structlog>=23.2.0
watchdog>=3.0.0
